% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tte.R
\name{generate_tte}
\alias{generate_tte}
\title{Simulate stratified time-to-event outcome randomized trial}
\usage{
generate_tte(x, block, ...)
}
\arguments{
\item{x}{the object}

\item{...}{Additional arguments (not used).}
}
\value{
TBA
}
\description{
Simulate stratified time-to-event outcome randomized trial
}
