% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{input_dropout}
\alias{input_dropout}
\title{Input dropout rate}
\usage{
input_dropout(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (not used).}
}
\value{
TBA
}
\description{
Input dropout rate
}
