% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{input_enroll}
\alias{input_enroll}
\title{Input enrollment rate}
\usage{
input_enroll(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{Additional arguments (not used).}
}
\value{
TBA
}
\description{
Input enrollment rate
}
