% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{input_failure}
\alias{input_failure}
\title{Input failure rate}
\usage{
input_failure(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments (not used).}
}
\value{
TBA
}
\description{
Input failure rate
}
