% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygsd-package.R
\docType{package}
\name{tidygsd-package}
\alias{tidygsd}
\alias{tidygsd-package}
\title{tidygsd: Group sequential design with tidy programming grammar.}
\description{
The goal of 'tidygsd' is to enable better grammar in group sequential designs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/littlebeannie/tidygsd}
  \item Report bugs at \url{https://github.com/littlebeannie/tidygsd/issues}
}

}
\author{
\strong{Maintainer}: Yujie Zhao \email{yujie.zhao@merck.com}

Authors:
\itemize{
  \item Keaven Anderson \email{keaven_anderson@merck.com}
  \item Nan Xiao \email{nan.xiao1@merck.com}
  \item Yilong Zhang \email{elong0527@gmail.com}
}

Other contributors:
\itemize{
  \item Hongtao Zhang \email{hongtao.zhang1@merck.com} [contributor]
  \item Larry Leon \email{larry.leon2@merck.com} [contributor]
  \item Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder]
}

}
\keyword{internal}
